grid 오브젝트를 color background 전용으로 만들고 앞으로 TectureGridObject 클래스를 만들 계획. 앞으로 오브젝트 회전에 따른 onTouch와 grid내의 오브젝트 포지션도 조정해야할듯.

+++ graphic2d/src/main/java/com/khgkjg12/graphic2d/RectObject.java
@@ -5,15 +5,15 @@ public class RectObject extends Object {
     private int mColor;
     int mWidth, mHeight;
 
-    public RectObject(int color, int width, int height, int z, int x, int y, String id){
+    public RectObject(int color, int width, int height, float z, int x, int y, String id){
         this(color, width, height, z, x, y, 0, 0, true, true, id);
     }
 
-    public RectObject(int color, int width, int height, int z, int x, int y, int degreeH, int degreeV, String id){
+    public RectObject(int color, int width, int height, float z, int x, int y, int degreeH, int degreeV, String id){
         this(color, width, height, z, x, y, degreeH, degreeV, true, true, id);
     }
 
-    public RectObject(int color, int width, int height, int z, int x, int y, int degreeH, int degreeV, boolean visibility, boolean clickable, String id){
+    public RectObject(int color, int width, int height, float z, int x, int y, int degreeH, int degreeV, boolean visibility, boolean clickable, String id){
         super(z, x, y, degreeH, degreeV, visibility, clickable, id);
         mColor = color;
         mWidth = width;
